gui: Add support for session autosaving
Currently if Builder crashes then all the pages opening/closing/moving
and any state a page had modified will be lost entirely, because saving
is only done when closing the primary workspace, which obviously
doesn't happen cleanly when it crashes.

So to avoid this big issue, this commit adds support for automatically
saving the grid's session state whenever a page is added/closed/moved,
or one of the worthy page properties changes, as defined by its session
addin.

So now Builder should be more reliable to such crashes, or at least
at least the last five minutes of work on the primary workspace will
be saved. It makes crashes less disruptive when working on a project.