gui: Remove redundant Open File "New document" menu button item
I noticed while adding the mnemonics that there was a second "Open file"
menu item, just that libdazzle de-duplicated them but it was visible
once I added a mnemonic to one of the two items, as they would no longer
match label.

Since the Open File menu item is already added in the editor plugin, for
both primary and editor workspace, this code is useless, so remove it.