tree: Set popover direction on initialization instead of mouse click popup
If the first time the popover is shown only the menu key is used, the popover
will be going upwards instead of to the right. So this commit instead sets
the popover position (i.e. direction) when initializing it instead of when
poping up the popover with the mouse (since the menu key doesn't go through
the same code path here).