plugin: deviced: Add run button support
When a deviced device is selected, clicking Run will now run the app on
the device.

Modified IdeDeployStrategy to be able to create a runner for the device,
and modified IdeRunManager to override the runtime's runner with one for
the device, if necessary.

There are a number of things that could be improved--the project is
reinstalled every time you click Run, and transferring flatpak commits
by delta is not supported--but it works well enough to use.