flatpak/daemon: Wire tests in meson
They are commented out for the time being because they don't pass and
I don't want nor know enough the daemon to fix it, and don't want to
block the MR with it until then.
They are useful to make sure the flatpak stuff doesn't break, or at
least is known early.