devhelp: Fix Ctrl+F shortcut by using an action shortcut
We can't use gtk_binding_entry_add_signal() because it requires us
to have the end focus (like an entry or button has) with can-focus=True,
which we can't have because we're just a container and not a "primitive"
widget. I'm not sure if that shortcut ever worked, and how though.

So instead use an action shortcut, and get back the search bar ;)