ide-marked-view: Escape text during markdown parsing
Before this commit, text sections of parsed markdown were not properly
escaped, e.g. an ampersand (&) in a text would end up in the final pango markup
unescaped, causing pango to fail parsing the markup.

With this commit, all leaf parts of the markdown (those that end up in the pango markup)
are now escaped.