flatpak: connect system/user fonts when running application
When running Flatpak applications, we are using 'flatpak build' since the
application has not yet been commited to an OSTree repository. That means
we do not get help from the system for things like "flatpak run" where
font caches would be wired into the process.

This attempts to alleviate that by manually doing the work that flatpak-run
does. It's not exact, and in particular we might risk writing to font
caches, but that appears to be okay.

Long term, we probably want to address this in:

  https://github.com/flatpak/flatpak/issues/4404

Fixes #1519