flatpak: disable related-refs on install/update
These can potentially break the install of SDKs if it pulls in something
with extra_data (like some FreeDesktop extensions). Of course, if any
particular SDK has extra_data, we'll fail the whole flatpak-bwrap dance
so in the long term, we probably need coordination from the Flatpak
Development D-Bus service if we want to keep going down this route.

Related #1532