intelephense: Bundle intelephense language server
We ran into a similar issue as the typescript language server plugin
that these node based language servers use the `processId` of the client
to check if the client is still alive. However because the flatpak build
of gnome builder lives in a different process namespace than intelephense when
it runs on the host this check to see if the client is still alive would
fail.

To solve this we include intelephense as a dependency for the flatpak
build similar to the typescript language server plugin.