flatpak: dont use IdeObject for flatpak client
Our client helper (which spawns the worker process) does not need to be
connected to the IdeObject tree for the project. This also helps ensure
that we do not run multiple copies of the gnome-builder-flatpak daemon
per Builder process.

Furthermore, it will allow us to spin up the process sooner at startup
so that there is less slowdown once opening a project.