flatpak: use flatpak_installation_list_remote_refs_sync()
Instead of calling flatpak_installation_update_remote_sync() at startup
we need to use flatpak_installation_list_remote_refs_sync() to help ensure
that we actually have refs that we can use when performing cache-only
lookups from the UI process. Otherwise, things like IsKnown() will fail.
That is a pretty bad look on a fresh install, so this will help with that.

It would be nice if we could do a bit less network I/O here but we were
already doing it so this at least fixes things.