flatpak: override some settings in flatpak environment
We want to make sure that applications can use things like the GTK
inspector out of the box. This makes it a bit easier by writing to the
keyfile that will be used by the flatpak application.

In the future, we want a more generic way to handle this so that we can
provide additional features in the Run popover such as theme.

Fixes #1264