app: work around ibus breakage
Ibus sort of screwed us over in that changing their keybinding means that
the Shell is stealing our input before we can ever get to it. This tries
to catch the key-release and do what our keybinding was meant to do.

There is one slight annoyance in that we get the emoji 'e' for just a
moment until the widget changes focus and the precommit text is altered.
A small price to pay to have things still work.

This will all go away in a GTK 4 version of Builder since we wont need
the snooper and can instead use a different keybinding for the new
integrated search/command-bar.

Fixes #1550