gopls: port to using IdeLspService
There is still one thing that is slightly missing from this, which is to
find the binary in ~/go/bin/gopls. I think now that various distributions
have gopls as packages, we can rely on that.

To improve this further (definitely yell if you use this!) we could add
alternate search paths in IdeLspService.