runtime: use user shell for locating program with which
This has to use the API with some special handling because we don't yet
have access to the libide-terminal (it requires libide-gui first).

It tries to use something like `bash --login -c 'which $program'` so that
we have the potential to discover where the binary is, even with an
alternate $PATH set by the user's shell.

Related #1384