clangd: add (disabled) clangd plugin
We already bundle clangd, so we might as well have the binding layer to be
able to use it. Currently this is disabled by default, but we can look at
enabling it if/when we are comfortable with how it works and if it can
load all our project build system bits without issue. In particular we need
to know it will support meson well (and that should be possible with the
compile_commands.json as long as it discovers it).