Revert "greeter: Use GtkFileChooserNative instead of GtkFileChooserDialog"
This reverts commit 85aee7693da257f4346799a5443021375502cd69.

This needs to be reverted because when used, we are not able to select
files matching the file-format if/when directory inode listings are
used.

Therefore, if you select "Meson (meson.build)" from the combo box in a
directory with a "meson.build" file, you can't actually select the
"meson.build" file at all.