handy: use HdyApplicationWindow
Now that we link against libhandy, we can use HdyApplicationWindow if we
fix a few pieces to go with it like the headerbar.

The one thing we loose for right now is the fullscreen button, but that
can come back if people really want it. It doesn't seem to be used much
right now.