foundry: rediagnose after pipeline loads
When restoring a document as we reload a project, we have a situation where
the build-flags will not yet be available. They won't be available until
the build has advanced or the pipeline finishes loading.

Previously, we would diagnose with an empty set of build flags, so things
like `#include "config.h"` would break. We still have that initially, but
now as soon as the pipeline finishes loading (or reloading after build
preference changes), those will go away without user interaction.