meson-template: Remove hard-coded default size
The default size of the application window should only be set in the
UI definition files. This way we can easily modify the generated
application in order to store and restore the window size and state.
If we pass default-width and default-height to the constructor of
the application window, we won't be able to set up a GSettings binding
in out instance initialization function.