intelephense & gvls: Do not override 'destroy'
Using `IdeObject.do_destroy` inside the plugins
replaces the Builder's default destroy behavior
(`ide_object_real_destroy`), which causes an
infinite loop when closing projects. With this
change, both plugins connect to the _destroy_
signal instead.

This also removes `do_parent_set` of GVls plugin
as it doesn't do anything.

Finally, GVls was attempting to close the
IdeLspClient in its destroy method, which was
unnecessary, as it can close itself. So, that
part is also removed.