completion: Fix completion for php variables
ide_completion_context_get_bounds does not include `$` in the bounds so
when we delete (begin, end) from the text buffer so we can insert the
text provided by the completion we don't remove the `$`. For php this is
problematic because the completion includes the `$` prefix, so when we
insert the completion we end up with 2 `$`.

For example `$t` triggers autocomplete with a proposal `$tree`, we apply the
proposal and end up with `$$tree`.

This patch also removes the common prefix of the completion item and the
`begin` GtkTextIter.

Fixes #1641