vls: enable search provider when build system supports vala
The goal here is to ensure we are less likely to spawn language servers
unless there is actually a need for the language server. For now, most
of these can be determined by the plugin type. However for search providers
they are enabled when the user types into the search box and begins typing.

The last thing we want to do while starting interactive search is:

 1) spawn new processes that can take away interactivity
 2) spawn a language server that has nothing to do with the project