libide/gui: start on new preferences window
The goal for this long term is to have a preferences window which can
look more like the Control Center in terms of how it works but support
both application preferences as well as project properties. They will
be displayed at separate times, through separate flows however. Some
settings which can have project overrides will be displayed redundantly
in the project properties window.