libide/gtk: add libide-gtk as landing spot for widgetry
We have a bunch of widgetry that might land in Builder again instead of
from libdazzle and this will be our landing spot for it along with things
like animations (until we can use those from Adwaita), custom buttons,
popovers, menu managers, etc.

The goal here is to get a lot more out of libide-gui and into libide-gtk
where we do not need to depend on much more than libide-core, io,
threading, etc.