libide/gui: add IdeWorkspaceAddin.focus_changed
This is to be used when the focus page changes within a workspace now that
the workspaces can have a bit more control without the need to indirect
through an IdeSurface.

A secondary goal here is to remove the need for IdeEditorAddin and just
use the IdeWorkspaceAddin directly (or IdeWorkbenchAddin when needed).