plugins/editorui: use workspace addin to add statusbar widgets
The goal here is for the statusbar widgets related to the editor to be
shown based on page changes. We still need actions for everything and
also need to be able to ignore focus changes where the widget has not
been destroyed or clicking on the menus will clear them.