snippet: Use single quotes in JS
gjs mandates single quotes for strings, which is the closest we get
to a canonical "GNOME JS" style, in particular as it is the base for
the style guidelines of gnome-shell and friends. As good as it would
be to have an option for the quoting style of string literals, it
makes more sense to follow the gjs style for now than using double
quotes, so adjust the snippets accordingly.

https://gitlab.gnome.org/GNOME/gnome-builder/-/issues/738