libide/threading: make IdeTask a final type
We originally thought we might subclass IdeTask to do specific work, but
that didn't pan out and there are better patterns for that rather than
subclassing the task itself.

So might as well make the type final and remove our priv indirection as
it makes it easier to debug from gdb too.