libide/code: defer boxed object release to idle
We don't want to do this synchronously as causing the box to be released
will result in g_object_run_dispose() being called which will disconnect
all the signal handlers.

That makes things difficult for plugins which might be racing against this
situation and cannot be semmetric with add/remove of signals.