libide/gui: improve panel positioning
This tries a bit harder to place the panels in positions that matter based
on column/row/depth.

Presumably, we'll need to alter things for the pages as their positions
will alter during runtime/switching pages but we can deal with that a bit
later when we have something to work with.