libide/code: fix implicit newline handling
This fixes a rather old bug with handling newlines related to change
monitors and how the implicit-trailing-newline gets applied.

They default to true, so we need to alter the buffer contents when setting
the property so that things stay in relatively same "shape" as they are
expected to go to/from disk.