libide/greeter: use flat for greeter headerbar
The greeter will likely go through a bunch more changes, but we want to
make it look less like the rest of the application and more of a landing
page which is temporary. A flat style might be one way to do that.

Longer term, I'd also like to rethink the flow between greeter pages
since they need so much surgery anyway.