libide/sourceview: handle request to scroll to insert
Generally we don't want to use this API, as it forces all views to scroll
to the insert location which are observing the buffer. However, it is used
for things like "restore-cursor" at load time, so we should at least start
with it working even if we remove it later.