libide/code: add IdeBuffer commit hooks
This is intended to help us manage operations with text "commit" operations
where the underlying buffer is changed. It doesn't give much information,
but then again, it's intended to be rather fast and correct in the sense
of requiring consumers to not do further modifications from their
function callbacks.

This can be used by spellcheck to update the region b+tree in concert with
the text buffer.