libide/gui: add observation helpers for panes
This can be useful so we can track panes without all the manual signal
connection wiring. We will still use signals for that with the new
destroyed signal on PanelWidget, but the helpers mean we are more likely
to get the tracking right in plugins where the panels might need to be
released or could have otherwise already been released.