libide/sourceview: add hover providers
This also cleans up the initialization and moves addin management to an
external source file to keep ide-source-view.c less busy.

To prevent things from loading until a language has been set, the default
value is "--disabled--" for the language so we are less likely to load
something incorrectly when the page is created.