plugins/symbol-tree: port symbol scope to GTK 4
This also moves the symbol-tree position to the statusbar of the
workspace rather than in the header of the frame. That might make things
easier for us to transition to tabs by default, as it gets one more thing
out of the frame header.

There is still more to do here, such as port the symbol tree to a
GtkListView and place it in a popover w/ search.