plugins/podman: use separate parser for layers
Also fix the string lifetime issue with out finding the layer was supposed
to work (by freeing caller to be replaced with return value). We need to
use a separate parser for the layer because it would cause pointers to
const strings to become invalid by clearing the old parser state.

Always set g_autoptr() values to NULL.