libide/gui: load resources after loading module
Otherwise, the resources could get loaded in the wrong order with regards
to menu merging (which is why we had the hack in place before to work
around it).

Since we don't need resources when the plugins load, but when their addins
are instantiated, this is fine and allows the resources to be dynamically
injected after the plugin, but before the addins.

It fixes the buildui plugin getting it's resources injected before the
project-tree has injected it's own menus.