libide/gui: create GtkShortcut from keybindings bundles
This still needs testing as it has only been compile tested. But the goal
is that we will parse JSON files of keybinding descriptions into a bundle
(which is a GListModel of GtkShortcut). We use the template-glib based
TmplExpr for "when" to determine if we can activate the
GtkShortcutAction from a GtkCallbackAction.

The GtkCallbackAction's user_data contains the phase, which we might want
to use at some point to get filtered GListModel by propagation phase for
the GtkShortcutController.