plugins/buildui: prototype select build target dialog
This isn't finished, but I prefer to get things in early and iterate. It
is just a window that we'll show to allow the user to select the default
build target.

It probably needs some whizbangs to show while it extracts content in case
that operation is slow (ie: requires building or configuring some stuff).

Also, we might want to allow a free-form target to be specified if the
backend supports it.