libide/foundry: take default-build-target into account
If the build manager has a default build target and a regular build
request was made, we should try to resolve the default build target and
provide it to the build pipeline.

We ignore this on things like install as that target will need things to
be built anyway and so a target would need to be ignored.