plugins/debuggerui: menu actions to break on warning/critical
This allows setting the G_DEBUG=fatal-warnings or G_DEBUG=fatal-criticals
environment variables with check menu items in the run menu. A new
"Debugger" sections section is added to contain this.

Even though we show these as checks, it's mostly to just not have to add
a third "Don't break on log" menu item as we can't really toggle them
individually.

This also makes the IdeDebugManager a GActionGroup like we do for other
manager objects with two new GAction for the breakpoint settings.