plugins/find-other-file: port to GTK 4
This uses a new design for the find-other-file which places the menu
button in the statusbar with a popover that is displayed when activated.

Plugins can further extend the list of files. In the future we may also
want to have a way to perform certain actions on the similar files such
as open the current file in a designer instead of as a source file.

But that can come later.