libide/greeter: toggle transparent separator on scroll
This makes an invisible separator visible when the scrolledwindow is
scrolled away from the zero position.

This uses a separator instead of box-shadow/border so that we can be
certain we don't affect the layout/sizing during the transition to reduce
chances of stutter.

Normally I don't use opacity: if i can help it, but in this case the GL
renderer knows how to simply alter alpha in the destination color instead
of an offscreen.

Fixes #1674