plugins/find-other-file: add model for directories
This is meant to be a model we can join with the existing glob results so
that we have all of the directories from the root to the target in the
result set.

Later on, we can make these directories be browsable instead of just
opening a directory list view.