plugins/mono: port to C
This is one of those types of plugins it would be really nice to simplify
so we don't have to create an addin subclass directly, or rather, it would
be nice if it could be done with XMACROs or something.

I still want PyGObject out of the UI process, but it would be nicer to
have something smaller than this.