plugins/buildui: add warning/error status indicator
This is currently placed in a menu button which does nothing, but the next
step here is to add a popover with various display items. That might allow
us to remove the info from the omnibar at some point.

More importantly, I want to make the process of "show issues, jump to
next warning/error" a bit more keyboard friendly w/o moving panels around.