plugins/gjs-symbols: remove gjs-symbols from GTK 4 port
This plugin doesn't appear to actually work anymore.

That is related to things beyond the expertise of this Builder maintainer.
We're looking to push some of that effort off to either GJS or another
tool for which we can integrate.

See https://gitlab.gnome.org/GNOME/gjs/-/issues/482 for more info.