plugins/gvls: drop gvls plugin
This module has not been maintained up to our necessary requirements with
moving towards IdeLspSevice and is far too much for us to maintain
voluntarily as we remove PyGObject from all bundled Builder plugins.

Any replacement will need to be written in C and most likely live on as
an external plugin as we already have a Vala language server which is
getting lots of maintenance upstream and significantly lower memory
usage/etc when used from Builder.

I don't like being the one in the position to make choices, but it seems
like the choice has already been made and I don't want the maintenance
burden falling on to me, as I've got enough things to do already.

We should, however, prioritize external plugins from our Flatpak and put
effort towards making that successful.