plugins/create-project: stub out new design
This doesn't implement everything yet, but it gets some of the machinery
in place. In particular, we don't yet have capabilities to move the
buttons up to the header bar for cancel/create and I'm not sure we'll
take the time to do that given the necessary bits to wire up actions too.

But once we have action redirection in GTK that'd be a lot easier.

The next step is to move the plugins into libide-projects so that the
template input object is most of the machinery.