libide/core: move ide_get_projects_dir() to libide-core
We can keep the implementation in libide-projects, but things need access
to it that may not be able to depend on libide-projects (due to cyclic
dependencies).

But libide-* was never mean't to be treated as a different library, just
different sub-systems within a single library. So we can just move the
declaration into libide-core's globals.