plugins/git: implement listing remote branches
Particularly from the IdeVcsCloner interface, but we could use this
elsewhere should we need to.

Note that by using this interface, the git plugin will create a
gnome-builder-git daemon instance for the IdeContext. Not really a big
deal in most cases, but might be strange if you tried to list branches,
then wen't back and opened some project without git vcs in that the
process will linger until the context ends.