libide/vcs: add in IdeVcsCloneRequest object
The goal here is to simplify cloning a repo into a helper object like we
did for creating project templates. It's really a "controller", if you
will, for the clone flow.

There is still additional work to do the actual clone, and I'd like to
add some busy state here for tracking progress with branch loading.

Additionally, we'll probably want to plumb a PTY through for clone
operations so subprocesses have access to that instead of shuttling
strings to us w/o any shell/control/etc.