plugins/web-browser: add simple web browser plugin
This is part of the work to start splitting up functionality from various
aspects of the html-preview plugin. We want it split into various plugins
rather than one which gets a bit hard to track what's going on.

This just adds a very simple way to create a new IdeWebkitPage separate
from all of the preview functionality. Primarily to test the widget before
porting the rest over.