plugins/sphinx-preview: add preview for reStructuredText
When using .rst without a conf.py, we will assume it is reStructuredText
without Sphinx. This uses a subprocess to parse the rst and convert it to
HTML by sending a Python script to standard input and the buffer contents
in a memfd (or anonymous fd) to the script running under python3.

While it requires spawning a process, I do prefer this to running Python's
docutils.core.publish_file() in process as it had some overrides of
builtins.open() which are undesireable. Additionally, it keeps the memory
fragmentation in the subprocess which can be torn down quickly.