plugins/meson-templates: start porting to C
This is a partial port of the meson-templates plugin to C. So far it only
includes the GNOME application template, but it gets things going for how
we'll go about porting the rest of them.

There is a bunch still to do here, but the C one works. I'd also like to
be able to automate testing of all of these at some point, or at least
that they expand.

But that will also come with cleaning up what they generate as I'm not
happy with how haphazard things have become.

The python plugin still is around, until we finish porting, then we can
remove it too.

As we don't need pattern icons anymore, those are removed.