plugins/meson-templates: remove legacy GTK 3 projects
We don't want to be encouraging the creation of GTK 3 projects at this
point, so just drop the crufty templates.

That means we don't have GTK templates for C++ or C♯ at this point, and
that is probably a good thing given their current state. If you'd like
to fix this, please do!