plugins/valac: replace vala-pack with valac plugin
This removes the vala-pack plugin which combined a pipeline error regex
with an indenter. Since indenters will come from GtkSourceView in the
GTK 4 port, we can just drop the indenter and focus on the error regex.

That can be a more specialized "valac" plugin similar to the mono plugin
for mcs/gmcs error output.