libide/io: improve icon override detection
The fallbacks we get from GThemedIcon are not necessarily in an order that
we might expect in priority. This overrides some detection so that we can
be more likely that we load our bundled override icon.

This fixes a number of bundled icons and adds some overrides for more.

 * Sometimes .md files might show up as musedata if you have applications
   installed with Wine which might override that.
 * Dockerfile now shows up like other build files we care about.
 * Ruby now has the proper icon
 * Python now has the proper icon
 * JavaScript and JSON both share the proper javascript icon
 * SQL now has the proper icon

Fixes #1684