plugins/shellcmd: add locality selection to dialog
We still need to wire this up, but it gives us the ability to actually
select a locality. The goal is to add this to shellcmd run command so that
we can use the appropriate terminal launcher.

These will likely be ignored when using a run command as a run target,
but that is to be expected.