libide/foundry: get stop signal from gsettings
This implements the functionality portion of being able to select the
stop signal to send an application to exit. Generally we want to just
use SIGKILL to be sure it's gone, but in some cases it might be nice to
be able to shutdown gracefully, at the risk of it not actually stopping
in a timely manner.

To handle that, a secondary call to force_quit will ignore the stop
signal gsetting.

There isn't a way to configure this from settings yet, but it does need
to be set for the per-project GSetting schema path, which makes it a
bit annoying to control manually at the moment.

To make this functionality somewhat useful, a preferences toggle will still
be needed in the "Configure Project" settings window.

Related #1685