libide/gui: add choice preference for stop signal
Continuation of #1685 to have a choice setting for setting the signal to
send the app when stoping it.

I think we still have some work to do to ensure that this is how the
application is getting stopped (as opposed to cancelling), but this makes
it somewhat testable now.

Related: 1685