libide/threading: add IdeUnixFDMap
This object is somewhat like GUnixFDList but meant to be for mapping a FD
in the current process into a FD number in a destination process. It can
be used by various subsystems (once ported off their private
implementations) by having it in libide-threading.

It includes a number of helpful features to make implementing things in
various dark corners of Builder easier. In particular it can handle all
the forms of get/steal/peek we need, opening files, and checking for
isatty() on various FDs.