libide/threading: port subprocess launcher to IdeUnixFDMap
This ports the subprocess launching to using the new IdeUnixFDMap as a
first test to see how the API is working for us. Ideally this will flush
out any bugs quickly so we can use it for run contexts and other areas
where you find something like a "FDMap" struct.