libide/foundry: start on new run context design
This is a layered run context design where we will allow tooling at various
layers to inject things as we process the creation of the launcher.

The goal is to simplify and/or remove IdeRunner completely in favor of this
and unify how the tooling injects/processes bits at each layer.