libide/foundry: couple deploy strategy to pipelines
This also adds a new local deploy strategy so that we can always go
through a deploy strategy for running. That can simplify divergence in
how things run, even if it's just pass-through.

Once this is in place, we can start looking at replacing the whole
IdeRunner infrastructure with something simpler.