libide/foundry: cleanup run_async implementation
We can clean a lot of this up by using run/finish pairs internally now that
we can rely on device-manager to do deploy-strategy for us and what not.

We still need to remove IdeRunner use, but this at least gets things into
shape where that is a lot easier to do.

We also start tracking the IdeRunCommand as part of the process even if we
are not using it yet.