plugins/deviced: implement basic command proxy
This isn't perfect by any means, but it does get us a situation where we
can probably be able to switch the deviced plugin to using IdeRunContext
as part of the revamp. It just is a helper program that can proxy some
signals across and tries to cleanup/etc after the program cleanly.