libide: Fix wrong file opened from build output
Builder offers the possibility to open a file from clicking a
warning/error message in the build output. However, when the path
clicked is relative and there is a file with the same name in the
project's source directory this file is opened, even if there is a
better match for the full path in the build directory.

This commit fixes such instances by looking for exact matches of the
full path in all the search roots (source directory and build directory)
before resorting to looking at the filename only.

Fixes #1687