libide/foundry: give pipeline access to runtime
The runtime may need information from the build pipeline to be able to
setup the environment for the run context. Rather than having it look that
up from the "manager" objects, give it to the runtime directly to avoid
any chances they get the wrong item back during suprious changes.