libide/threading: add stream helper for IdeUnixFDMap
This helps create a GIOStream that can be used from the Builder side to
a pair of pipes that are delivered to the subprocess. This is generally
used to control a subprocess without having to go through all the normal
bits of GSubprocess io-stream management (which may not be possible given
API constraints).