libide/foundry: port plugins to IdeRunTool
This is the first step in moving away from the IdeRunHandler, and another
step towards removing IdeRunner. I notice some things still wrong with
regular running, debugging and what-not seems to work.

Sysprof is still unported, but this moves us in a direction that is much
nicer for handling that too.