plugins/flatpak: give flatpak-build access to minimal environment
Otherwise, it won't setup various things for us in the app context. This
isn't required everywhere, so the environment is set again at a higher
level from here when running.

We only set this in the "flatpak build" environment for *running* because
we don't want it available in builds since flatpak-builder wouldn't have
access to that either.