libide/foundry: simplify base implementation to check for program
This, so far, should almost never get used, so just provide a reasonable
fallback for new implementations. Runtime should override this to use their
run contexts as appropriate (or something faster when possible).

The GbpHostRuntime already handles the special case above so we can just
go ahead and delete that case.