plugins/testui: bind model from test-manager to listview
We still need something to kick off the loading of the run commands to be
filtered by the test-manager though.

What we might want to do, is to have the IdeRunManager maintain that core
model and have it update when building. At that point, it can always hand
out the same model for results.