libide/foundry: add IdeRunCommands
This is the start of an abstraction to simplify getting access to run
commands and to make them persistent/shared across all the subsystems.

It's still incomplete, but shows what the API will look like. I want to
hook some things up so that we wait until the pipeline is ready before we
populate providers.