libide/foundry: allow setting locality for unit tests
This allows running the unit tests in either the runtime or pipeline
environments. This can be a bit more convenient than spinning up a
fake display for unit tests.

We do want peope to be aware that unit tests should be automated in the
CI, and therefore display access is not necessarily a given.