build: reset version macros for 43
We are switching GTK versions, so it's a great time to remove all of our
previous versioning information for the new release series. We are still
ABI bound per-release, much like other GNOME modules, but this just cleans
things since any external plugins would have to be rewritten anyway with
the GTK 4 change.