libide-sourceview: port to GtkSourceView 5.x
We can delete a lot of bits from libide-sourceview as it has been moved
upstream to GtkSourceView in various incarnations.

Not everything here is coming back, at least immediately. Multiple cursors
is one bit that was never implemented particularly robustly and will need
further work if it is to come back as part of the GTK 4 port.