libide-gui: port preferences to GTK 4
This provides a new preferences abstraction. I still don't love it, but it
appears to work for the time being. We have a lot more preferences still
to port over though so help with that is appreciated.

The preferences window can be in one of three modes (only two are used)
which is app or project mode. That determines what types of things are
added to the preferences window.